Set the type attribute of the buttons for logging out of all sessions to buttons
By not setting this attribute, the buttons default to submit and as the first
submit button on the page, it causes pressing the enter key to log you out of
all sessions. This change restores the pre 4.1 behavior where pressing enter
while focused on a form field submits the form.

props ocean90
fixes #30871 for trunk



git-svn-id: http://develop.svn.wordpress.org/trunk@31010 602fd350-edb4-49c9-b593-d223f7449a82