Multisite: Use single URL input when editing sites in a subdirectory configuration
In [32759], we adjusted `site-info.php` to display a single input for a site's full URL if the network was configured for subdomains. We also enforced path only editing for non-subdomain networks, which is a regression in expected behavior.

The full URL of a site can now be edited in both subdomain and subdirectory configurations.

Props @michaelryanmcneill.
Fixes #22383.


git-svn-id: http://develop.svn.wordpress.org/trunk@33586 602fd350-edb4-49c9-b593-d223f7449a82