Don't show the "Add New" link on `post-new.php`.
Using the same `$post_new_file` global variable name in `wp-admin/menu.php` in [33723] was unintentional, but we're already unsetting other variables that are in global scope so we can just unset this one as well. Someday the admin menu won't be a big pile of sad and cryptic globals. Maybe.

props johnjamesjacoby, norcross.
fixes #34045.


git-svn-id: http://develop.svn.wordpress.org/trunk@34723 602fd350-edb4-49c9-b593-d223f7449a82