Introduce a language chooser to the site signup process on Multisite.
If your Multisite installation is one of the relatively few that allows new sites to be registered, either by existing users or by new visitors to your site, a 'Site Language' dropdown menu will now be presented if your network has additional languages installed. This option defaults to the value of the 'Default Language' setting on the Network Admin Settings screen, and is restricted to currently installed languages.

The languages available in this setting can be controlled via the `signup_get_available_languages` filter. To disable it completely, return an empty array.

Fixes #33844
Props DrewAPicture, johnbillion


git-svn-id: http://develop.svn.wordpress.org/trunk@35152 602fd350-edb4-49c9-b593-d223f7449a82