Customizer: Prevent `nav_menu_item` settings from becoming dirty when their controls are set up.
Since `wp_setup_nav_menu_item()` returns the `classes` property as an array but the Customizer manages the value as a string, the setting needs to initially export the value as a string. This prevents the `classes` property type change from causing the setting to get marked as dirty even though nothing changed. This is a regression from [34788].

See #34111.


git-svn-id: http://develop.svn.wordpress.org/trunk@35308 602fd350-edb4-49c9-b593-d223f7449a82