Emoji: Ensure twemoji kicks in on certain DOM mutations.
Twemoji will replace the `img` with the emoji character, in the event that the image fails to load. We deliberately avoid trying to change that emoji character when it's changed back. We do need to replace emoji characters that are changed by something other than Twemoji, which this rectifies.

Fixes #34640.



git-svn-id: http://develop.svn.wordpress.org/trunk@35637 602fd350-edb4-49c9-b593-d223f7449a82