Themes: Show template loading error to users with `switch_themes` cap.
In [36338], a template loading error was shown only to users with the `install_themes` capability. This is now displayed users with the `switch_themes` capability, as users with this cap can at least switch to a different theme. Also, this will now show for site administrators in multisite, whereas `install_themes` is limited to superadmins.

Props dd32.
See #21931.


git-svn-id: http://develop.svn.wordpress.org/trunk@36344 602fd350-edb4-49c9-b593-d223f7449a82