Fix test related to cap check in `get_edit_term_link()`.
The test was introduced in [36646] was intended to demonstrate that the cap
check in `get_edit_term_link()` uses the custom capability from the correct
taxonomy when `$taxonomy` is not specified but must be inferred from the
provided term. However, the test was wrongly written in pretty much every way.
Antiprops boonebgorges.

Props swissspidy.
See #35922.

git-svn-id: http://develop.svn.wordpress.org/trunk@36986 602fd350-edb4-49c9-b593-d223f7449a82