Smilies: Move `convert_smilies` to happen later in the `the_content` filter.
In particular, we want it to occur after shortcode handling. The smiley conversion routine doesn't have any concept of shortcode structure, so may inadvertantly replace a smiley with HTML inside a shortcode attribute, which will cause the shortcode to not be parsed correctly.

Props Unyson for the initial suggested fix.

Fixes #36306. 



git-svn-id: http://develop.svn.wordpress.org/trunk@37298 602fd350-edb4-49c9-b593-d223f7449a82