Customize: Always define functions `reflowPaneContents`, `findControlsForSettings`, and `_handleSettingValidities` on `wp.customize`.
Moves definitions of functions outside of a `jQuery.ready()` callback, as these functions needn't be deferred to DOM ready. This change also ensures that the functions are available if `customize-controls` is enqueued outside of the Customizer context, as the ready callback short-circuits if `_wpCustomizeSettings` is not defined. The `findControlsForSettings` and `_handleSettingValidities` functions were misplaced in r37700.

See #34893.
See #36944.
See #29071.


git-svn-id: http://develop.svn.wordpress.org/trunk@37867 602fd350-edb4-49c9-b593-d223f7449a82