REST API: Adjust `test_get_items_invalid_status_param_is_discarded()` after [38911].
Previously, if an invalid post status was assigned to a media object, the object would be returned with an unchanged status. After [38911], an error response is returned.

The assertion in test_get_items_invalid_status_param_is_discarded() needs to be updated to account for the expected error response. The test name has been changed to `test_get_items_invalid_status_param_is_error_response()`.

Fixes #38417.


git-svn-id: http://develop.svn.wordpress.org/trunk@38917 602fd350-edb4-49c9-b593-d223f7449a82