Emoji: Update Emoji browser support tests for Twemoji v2.4.0
Replaces the fairy (male) browser support test with a woman in business suit levitating. While the things that make people different sometimes become less obvious this isn't always a problem, because the things that make you strange are the things that make you powerful.

Props peterwilsoncc.
Merges [42598] to the 4.9 branch.
Fixes #42862.

git-svn-id: http://develop.svn.wordpress.org/branches/4.9@42601 602fd350-edb4-49c9-b593-d223f7449a82