Customizer: In menus, reset results when closing the 'add items' panel.
Trigger a keyup event when clearing the search field in response to closing the add item panel. The keyup event triggers a search and resets the results. Previously, the search field was cleared while the potentially blank search results were left in place making it impossible to select new menu items.

Merge of [42744] to the 4.9 branch.

Props Blair jersyer, aranwer104, afercia.
See #43333.

git-svn-id: http://develop.svn.wordpress.org/branches/4.9@42846 602fd350-edb4-49c9-b593-d223f7449a82