REST API JavaScript Client: Support an empty string for `nonce` to disable sending the X-WP-Nonce header.
Passing a `nonce` argument with an empty string to `wp.api.init()` now does no longer fall back to `wpApiSettings.nonce`. This makes it possible to stop sending nonce headers, for example to a read-only endpoint on another site in a multisite install.

Merge of [42852] to the 4.9 branch.

Props adamsilverstein, FPCSJames, ocean90, swissspidy.
See #42948, #43266.

git-svn-id: http://develop.svn.wordpress.org/branches/4.9@42854 602fd350-edb4-49c9-b593-d223f7449a82