Privacy: Redirect to newly created Privacy Policy page to improve UX.
Previously the user was shown a message that the page was created, but might not understand that they still need to visit the page and publish it. Redirecting them to the page makes it more obvious that additional steps are involved.

Props Clorith, xkon, azaozz.
Merges [43160] to the 4.9 branch.
Fixes #43926.

git-svn-id: http://develop.svn.wordpress.org/branches/4.9@43161 602fd350-edb4-49c9-b593-d223f7449a82