Privacy: Return before scheduling cron during install to avoid error.
r43046 introduced wp_schedule_delete_old_privacy_export_files() to schedule the wp_privacy_delete_old_export_files cron job, but it did not check to make sure it wasn't running in the context of the install process. When it did run in that context, it created a database error, because the necessary database tables don't exist at that point.

Checking the current context and returning early during the installation phase avoids that issue.

Props helen, timothyblynjacobs, iandunn.
Merges [43162] to the 4.9 branch.
Fixes #43952.

git-svn-id: http://develop.svn.wordpress.org/branches/4.9@43163 602fd350-edb4-49c9-b593-d223f7449a82