Twenty Nineteen: Make sure links are followed on touchend, not touchstart.
The theme's original navigation JavaScript was making it so all links on a site were immediately followed on touchstart when using a touch-enabled device. This update makes sure links are followed at touchend, to improve usability and menu behavior.

Already committed to the 5.0 branch in [44357].

Props anevins, panchen, kjellr.
Fixes #45510.



git-svn-id: http://develop.svn.wordpress.org/trunk@44358 602fd350-edb4-49c9-b593-d223f7449a82