Customize: Safeguard a check on the `customize_validate_{$setting_id}` filter value to ensure it is a `WP_Error`.
While the filter is documented to only support a `WP_Error`, it has been a common practice to return true in a validation function if no errors have occurred. This was already caught when the same filter was executed in `WP_Customize_Setting`, it was however missing in `WP_Customize_Manager::validate_setting_values()`.

Props flixos90.

Merges [43578] to the 5.0 branch.
Fixes #44809.

git-svn-id: http://develop.svn.wordpress.org/branches/5.0@44392 602fd350-edb4-49c9-b593-d223f7449a82