Twenty Seventeen: Improve selectors for block editor custom colors.
Twenty Seventeen's original styles for the block editor custom colors had some issues: they weren't being applied to the button blocks due to lack of specificity, and when applied to paragraph blocks, there was no padding in the editor. This update makes sure the colors and related styles work as expected.

Fixes #45426.



git-svn-id: http://develop.svn.wordpress.org/trunk@44402 602fd350-edb4-49c9-b593-d223f7449a82