Accessibility: Improve buttons placement in the Comments Quick edit/Reply form.
Proximity of related information is an inclusive design principle that benefits everyone and it's particularly important for users with reduced visual field, low vision, or other vision or cognitive impairments. User interface controls that are logically grouped should be placed close to each other.
Also, this change makes the buttons visual order and the DOM order match.

Fixes #45972.


git-svn-id: http://develop.svn.wordpress.org/trunk@44757 602fd350-edb4-49c9-b593-d223f7449a82