Accessibility: Improve the image and gallery widgets preview accessibility.
In an authoring context, the image `alt` attribute purpose is different from the one for the front end. For example, screen reader users need to know what the selected image is, even when the original `alt` value is empty.

This change introduces a new pattern for the `alt` text in an authoring context:
- uses the `alt` text if not empty
- when there's no `alt` text, informs users the image has no alternative text and provides a reference to the image filename

Also, makes the gallery media widget preview an unordered list to make screen readers announce the number of images automatically.

Fixes #43137.


git-svn-id: http://develop.svn.wordpress.org/trunk@44767 602fd350-edb4-49c9-b593-d223f7449a82