Multisite: Ensure site meta caches are not primed unless the `wp_blogmeta` table is available.
Prior to this change, querying sites early in the bootstrap process could potentially cause a fatal error, since at that stage the filter to bail on updating site meta cache if the respective database table has not been installed yet is not hooked in yet. This changeset forces the filter to be added if that is not already the case.

Props spacedmonkey.
Fixes #46167.


git-svn-id: http://develop.svn.wordpress.org/trunk@44925 602fd350-edb4-49c9-b593-d223f7449a82