Privacy: Remove unnecessary `WP_Error` when handling `confirmaction` requests.
By reordering the logic when handling the `confirmaction` action in `wp-login.php`, the need for a new `WP_Error` object to be created can be eliminated. The error message can be passed directly into a `wp_die()` call, matching the other validation errors in related code.

Props garrett-eclipse, birgire.
Fixes #44901.

git-svn-id: http://develop.svn.wordpress.org/trunk@44931 602fd350-edb4-49c9-b593-d223f7449a82