Accessibility: Make sure the Site Health layout tables are correctly linearized.
Adds `role="presentation"` to the `<table>` elements used in the Site Health reports.

Ideally, HTML tables should be used for tabular data. When tables are used for layout purposes, it's important to remove any native semantics so that assistive technologies can correctly announce the table content in a linearized fashion.

Props chetan200891.
Fixes #46725.


git-svn-id: http://develop.svn.wordpress.org/trunk@45075 602fd350-edb4-49c9-b593-d223f7449a82