Bootstrap/Load: Only send the recovery mode email if the error occurs on a protected endpoint.
This reduces the potential email noise to places where the fatal error could be interfering with the login and admin experience.

Once the user is in recovery mode, any fatal errors (even if they aren't in a protected endpoint) are handled and the plugin or theme will be paused.

Props TimothyBlynJacobs.
See #46950.

git-svn-id: http://develop.svn.wordpress.org/trunk@45238 602fd350-edb4-49c9-b593-d223f7449a82