General: Update support forum links.
There are a lot of places in Core that link to https://wordpress.org/support/ for the support forums, but that's now the URL for HelpHub. The new forums link is https://wordpress.org/support/forums/.

This is a follow-up to [45140], which changed the links in help tabs.

Props dilipbheda.
Fixes #47239. See #46790.

git-svn-id: http://develop.svn.wordpress.org/trunk@45412 602fd350-edb4-49c9-b593-d223f7449a82