Taxonomy: Standardize docs for `wp_list_categories()`.
Many of the parameters historically documented on this function are merely
passed through to `get_categories()` and its underlying functions. In order to
avoid unnecessary duplication of function arguments, we remove documentation of
these parameters, and replace it with reference to the wrapped functions.

Parameters that receive special treatment or are otherwise changed before
being passed through continue to be documented as part of `wp_list_categories()`.

Props audrasjb, itowhid06, birgire.
Fixes #47896.

git-svn-id: http://develop.svn.wordpress.org/trunk@45894 602fd350-edb4-49c9-b593-d223f7449a82