Build/Test Tools: Fork and Update `grunt-replace`
The version of grunt replace that is bundled in core is using an outdated version of lodash that is bringing 2 low, 3 high, and 1 critical issue. This package is currently abandoned. There is a community forked version, but that is also harboring some similar security issues.

This switches to a fork by @whyisjake and causes no change to the build.

See #48203.
Fixes #48217.
Props whyisjake, netweb for testing.



git-svn-id: http://develop.svn.wordpress.org/trunk@46403 602fd350-edb4-49c9-b593-d223f7449a82