Pings/Trackbacks: Remove a failing test that won't pass as long as `WP_IMPORTING` gets set during tests.
The `_publish_post_hook()` function checks for `WP_IMPORTING` before setting meta fields fir enclosures and pings, which means this test is doomed to fail.

The test can be re-implemented if the `WP_IMPORTING` constant gets moved to a function similar to how `wp_installing()` works for the `WP_INSTALLING` constant.

See #48014


git-svn-id: http://develop.svn.wordpress.org/trunk@46428 602fd350-edb4-49c9-b593-d223f7449a82