Accessibility: Script Loader: Add jQuery as dependency of `admin-bar`.
On the front end, themes that don't use jQuery make the admin bar fallback to the non-jQuery implementation. Some important features miss from the non-jQuery `admin-bar.js` part, for example `hoverintent` and, most importantly, the touch events for the mobile menu don't work at all.

Enqueueing jQuery is the simplest option for now. For the future, a complete rewriting of `admin-bar.js` in plain JavaScript is highly recommended.

See #47069.


git-svn-id: http://develop.svn.wordpress.org/trunk@46440 602fd350-edb4-49c9-b593-d223f7449a82