Network Admin: Improve Site States in List Table rows.
This commit switches a comparison to a Yoda condition, performs a more strict `intval()` check on the value of the Site Status column, and prevents a specific Site State from appearing in List Table rows when filtering by that same Site State already (to match the behavior of other List Table State implementations.)

Fixes #37684. Props pbiron.

git-svn-id: http://develop.svn.wordpress.org/trunk@46441 602fd350-edb4-49c9-b593-d223f7449a82