Code Modernisation: Introduce the spread operator in `Walker::walk()` and `::paged_walk()`.
Rather than relying on `func_get_args()` to retrieve arbitrary function arguments, we can now use the spread operator to assign them directly to a variable.

This re-applies [45624] and reverts [45640], with a dev note upcoming for plugin authors to maintain backward compatibility with old versions of WordPress.

Props jrf.
See #47678.

git-svn-id: http://develop.svn.wordpress.org/trunk@46442 602fd350-edb4-49c9-b593-d223f7449a82